Logger API update
