+++ catalyst/loggers/console.py
@@ -1,7 +1,8 @@
-from typing import Dict
+from typing import Dict, TYPE_CHECKING
 
 from catalyst.core.logger import ILogger
-
+if TYPE_CHECKING:
+    from catalyst.core.runner import IRunner
 
 def _format_metrics(dct: Dict):
     return " | ".join([f"{k}: {float(dct[k])}" for k in sorted(dct.keys())])
@@ -22,7 +23,7 @@ class ConsoleLogger(ILogger):
         super().__init__(log_batch_metrics=False, log_epoch_metrics=True)
         self._log_hparams = log_hparams
 
-    def log_hparams(self, hparams: Dict) -> None:
+    def log_hparams(self, hparams: Dict, runner: IRunner = None) -> None:
         """Logs hyperparameters to the console."""
         if self._log_hparams:
             print(f"Hparams: {hparams}")
@@ -30,27 +31,17 @@ class ConsoleLogger(ILogger):
     def log_metrics(
         self,
         metrics: Dict[str, float],
-        scope: str = None,
-        # experiment info
-        num_epochs: int = 0,
-        epoch_step: int = 0,
-        batch_step: int = 0,
-        sample_step: int = 0,
-        # loader info
-        loader_key: str = None,
-        loader_batch_len: int = 0,
-        loader_sample_len: int = 0,
-        loader_batch_step: int = 0,
-        loader_sample_step: int = 0,
+        scope: str,
+        runner: IRunner,
     ) -> None:
         """Logs loader and epoch metrics to stdout."""
         if scope == "loader":
-            prefix = f"{loader_key} ({epoch_step}/{num_epochs}) "
+            prefix = f"{runner.loader_key} ({runner.epoch_step}/{runner.num_epochs}) "
             msg = prefix + _format_metrics(metrics)
             print(msg)
         elif scope == "epoch":
             # @TODO: remove trick to save pure epoch-based metrics, like lr/momentum
-            prefix = f"* Epoch ({epoch_step}/{num_epochs}) "
+            prefix = f"* Epoch ({runner.epoch_step}/{runner.num_epochs}) "
             msg = prefix + _format_metrics(metrics["_epoch_"])
             print(msg)
