+++ catalyst/loggers/wandb.py
@@ -1,4 +1,4 @@
-from typing import Dict, Optional
+from typing import Dict, Optional, TYPE_CHECKING
 import os
 import pickle
 
@@ -9,7 +9,8 @@ from catalyst.settings import SETTINGS
 
 if SETTINGS.wandb_required:
     import wandb
-
+if TYPE_CHECKING:
+    from catalyst.core.runner import IRunner
 
 class WandbLogger(ILogger):
     """Wandb logger for parameters, metrics, images and other artifacts.
@@ -97,18 +98,7 @@ class WandbLogger(ILogger):
         tag: str,
         artifact: object = None,
         path_to_artifact: str = None,
-        scope: str = None,
-        # experiment info
-        num_epochs: int = 0,
-        epoch_step: int = 0,
-        batch_step: int = 0,
-        sample_step: int = 0,
-        # loader info
-        loader_key: str = None,
-        loader_batch_len: int = 0,
-        loader_sample_len: int = 0,
-        loader_batch_step: int = 0,
-        loader_sample_step: int = 0,
+        runner: IRunner = None,
     ) -> None:
         """Logs artifact (arbitrary file like audio, video, weights) to the logger."""
         if artifact is None and path_to_artifact is None:
@@ -117,7 +107,7 @@ class WandbLogger(ILogger):
         artifact = wandb.Artifact(
             name=self.run.id + "_aritfacts",
             type="artifact",
-            metadata={"loader_key": loader_key, "scope": scope},
+            metadata={"loader_key": runner.loader_key},
         )
 
         if artifact:
@@ -137,59 +127,35 @@ class WandbLogger(ILogger):
         self,
         tag: str,
         image: np.ndarray,
-        scope: str = None,
-        # experiment info
-        num_epochs: int = 0,
-        epoch_step: int = 0,
-        batch_step: int = 0,
-        sample_step: int = 0,
-        # loader info
-        loader_key: str = None,
-        loader_batch_len: int = 0,
-        loader_sample_len: int = 0,
-        loader_batch_step: int = 0,
-        loader_sample_step: int = 0,
+        runner: IRunner = None,
     ) -> None:
         """Logs image to the logger."""
-        self.run.log(
-            {f"{tag}_scope_{scope}_epoch_{epoch_step}.png": wandb.Image(image)},
-            step=sample_step,
-        )
+        self.run.log({f"{tag}.png": wandb.Image(image)}, step=runner.sample_step)
 
-    def log_hparams(self, hparams: Dict) -> None:
+    def log_hparams(self, hparams: Dict, runner: IRunner = None) -> None:
         """Logs hyperparameters to the logger."""
         self.run.config.update(hparams)
 
     def log_metrics(
         self,
         metrics: Dict[str, float],
-        scope: str = None,
-        # experiment info
-        num_epochs: int = 0,
-        epoch_step: int = 0,
-        batch_step: int = 0,
-        sample_step: int = 0,
-        # loader info
-        loader_key: str = None,
-        loader_batch_len: int = 0,
-        loader_sample_len: int = 0,
-        loader_batch_step: int = 0,
-        loader_sample_step: int = 0,
+        scope: str,
+        runner: IRunner,
     ) -> None:
         """Logs batch and epoch metrics to wandb."""
         if scope == "batch" and self.log_batch_metrics:
             metrics = {k: float(v) for k, v in metrics.items()}
             self._log_metrics(
                 metrics=metrics,
-                step=sample_step,
-                loader_key=loader_key,
+                step=runner.sample_step,
+                loader_key=runner.loader_key,
                 prefix="batch",
             )
         elif scope == "loader" and self.log_epoch_metrics:
             self._log_metrics(
                 metrics=metrics,
-                step=sample_step,
-                loader_key=loader_key,
+                step=runner.sample_step,
+                loader_key=runner.loader_key,
                 prefix="epoch",
             )
         elif scope == "epoch" and self.log_epoch_metrics:
@@ -197,7 +163,7 @@ class WandbLogger(ILogger):
             per_loader_metrics = metrics[loader_key]
             self._log_metrics(
                 metrics=per_loader_metrics,
-                step=sample_step,
+                step=runner.sample_step,
                 loader_key=loader_key,
                 prefix="epoch",
             )
