+++ catalyst/runners/supervised.py
@@ -1,11 +1,27 @@
 from typing import Any, List, Mapping, Tuple, Union
+from collections import OrderedDict
 
 import torch
 
+from catalyst.callbacks.backward import BackwardCallback
+from catalyst.callbacks.criterion import CriterionCallback
+from catalyst.callbacks.optimizer import OptimizerCallback
+from catalyst.callbacks.scheduler import SchedulerCallback
+from catalyst.core.callback import (
+    Callback,
+    IBackwardCallback,
+    ICriterionCallback,
+    IOptimizerCallback,
+    ISchedulerCallback,
+)
+from catalyst.core.engine import IEngine
+from catalyst.core.misc import callback_isinstance
 from catalyst.core.runner import IRunner
+from catalyst.runners.runner import Runner
+from catalyst.typing import RunnerModel, TorchCriterion, TorchOptimizer, TorchScheduler
 
 
-class ISupervisedRunner(IRunner):
+class ISupervisedRunner(Runner):
     """IRunner for experiments with supervised model.
 
     Args:
@@ -146,4 +162,101 @@ class ISupervisedRunner(IRunner):
         self.batch = {**batch, **self.forward(batch)}
 
 
-__all__ = ["ISupervisedRunner"]
+class SupervisedRunner(ISupervisedRunner, Runner):
+    """Runner for experiments with supervised model.
+
+    Args:
+        model: Torch model instance
+        engine: IEngine instance
+        input_key: key in ``runner.batch`` dict mapping for model input
+        output_key: key for ``runner.batch`` to store model output
+        target_key: key in ``runner.batch`` dict mapping for target
+        loss_key: key for ``runner.batch_metrics`` to store criterion loss output
+
+    SupervisedRunner logic pseudocode:
+
+    .. code-block:: python
+
+        batch = {"input_key": tensor, "target_key": tensor}
+        output = model(batch["input_key"])
+        batch["output_key"] = output
+        loss = criterion(batch["output_key"], batch["target_key"])
+        batch_metrics["loss_key"] = loss
+
+    .. note::
+        Please follow the `minimal examples`_ sections for use cases.
+
+        .. _`minimal examples`: http://github.com/catalyst-team/catalyst#minimal-examples  # noqa: E501, W505
+    """
+
+    def __init__(
+        self,
+        model: RunnerModel = None,
+        engine: IEngine = None,
+        input_key: Any = "features",
+        output_key: Any = "logits",
+        target_key: str = "targets",
+        loss_key: str = "loss",
+    ):
+        """Init."""
+        ISupervisedRunner.__init__(
+            self,
+            input_key=input_key,
+            output_key=output_key,
+            target_key=target_key,
+            loss_key=loss_key,
+        )
+        Runner.__init__(self, model=model, engine=engine)
+
+    @torch.no_grad()
+    def predict_batch(self, batch: Mapping[str, Any], **kwargs) -> Mapping[str, Any]:
+        """
+        Run model inference on specified data batch.
+
+        .. warning::
+            You should not override this method.
+            If you need specific model call, override runner.forward() method.
+
+        Args:
+            batch: dictionary with data batch from DataLoader.
+            **kwargs: additional kwargs to pass to the model
+
+        Returns:
+            Mapping[str, Any]: model output dictionary
+        """
+        batch = self._process_batch(batch)
+        output = self.forward(batch, **kwargs)
+        return output
+
+    def get_callbacks(self) -> "OrderedDict[str, Callback]":
+        """Returns the callbacks for the experiment."""
+        callbacks = super().get_callbacks()
+        callback_exists = lambda callback_fn: any(
+            callback_isinstance(x, callback_fn) for x in callbacks.values()
+        )
+        if isinstance(self._criterion, TorchCriterion) and not callback_exists(
+            ICriterionCallback
+        ):
+            callbacks["_criterion"] = CriterionCallback(
+                input_key=self._output_key,
+                target_key=self._target_key,
+                metric_key=self._loss_key,
+            )
+        if isinstance(self._optimizer, TorchOptimizer) and not callback_exists(
+            IBackwardCallback
+        ):
+            callbacks["_backward"] = BackwardCallback(metric_key=self._loss_key)
+        if isinstance(self._optimizer, TorchOptimizer) and not callback_exists(
+            IOptimizerCallback
+        ):
+            callbacks["_optimizer"] = OptimizerCallback(metric_key=self._loss_key)
+        if isinstance(self._scheduler, TorchScheduler) and not callback_exists(
+            ISchedulerCallback
+        ):
+            callbacks["_scheduler"] = SchedulerCallback(
+                loader_key=self._valid_loader, metric_key=self._valid_metric
+            )
+        return callbacks
+
+
+__all__ = ["ISupervisedRunner", "SupervisedRunner"]
